Update heat_radius calculation in PeopleMovementHeatmap class

+++ people_move_heatmap.py
@@ -27,7 +27,7 @@ class PeopleMovementHeatmap:
             self.device = "cpu"
 
         self.heat_per_person = 10
-        self.heat_radius = 100 * scale
+        self.heat_radius = int(100 * scale)
         self.heat_per_distance = 2
         self.heat_decay = 0.9
         self.model = DetrForObjectDetection.from_pretrained("facebook/detr-resnet-50").to(self.device)
